Use $xpath->evaluate instead of $xpath->query
The results of `$xpath->query()` are always expected to be a Node
list. However, some of the only (the only?) uses of `EXTRACTVALUE`
were to evaluate `count()` expressions. Those will never return
a Node list. So `evaluate()` is used instead. If something that is
not an object is returned, we return that immediately. Otherwise
we assume it is a Node list.

Maybe this is an incorrect assumption but we can hopefully address
that when that comes up.

Initial discussion on implementation:

 o https://gist.github.com/8a79193b743287ed11e3