fix nested joins where clause generation
when having nested joins the current QOMWalking is falsly
assuming the right join as the most left join resulting in
a wrong query when having multiple joins with different
node_types.
included tests highlight the problem:
i have a document and another document thats a weak reference,
I then join the weakReference but also need to join a child document
if my document now has a type that the child doesnt have the resulting
query will falsly assume the root type is that of the weakReference because
its using that alias instead of the most left document.