prevent type error on Postgresql
Prevent a cast error on Postgresql when no target is found for the given
path. This by adding an early return of an empty array. MySQL and
possibly other databases would cast the false value to 0, after which
the query result would still be an enpty array as no node can have a
target_id equaling 0.